Switched to Rider, reformatting
