+++ project_docs/Scenario.md
@@ -2,26 +2,35 @@
 
 ## Morning
 
-Zoe wakes up feeling a bit off. She can’t quite pinpoint why, but the gloomy weather outside doesn’t help. As she gets ready for work, she wonders, “Have I been feeling this way a lot lately, or is it just today?”
+Zoe wakes up feeling a bit off. She can’t quite pinpoint why, but the gloomy weather outside doesn’t help. As she gets
+ready for work, she wonders, “Have I been feeling this way a lot lately, or is it just today?”
 
 She makes a mental note to reflect on it later but quickly forgets as she rushes out the door.
 
 ## Late Morning
 
-At work, Zoe finds herself snapping at a colleague over a minor issue. She immediately regrets it, realizing she’s been more irritable lately. “Maybe I’m just stressed,” she thinks. But how can she be sure? She wishes she had a way to keep track of her mood patterns over time.
+At work, Zoe finds herself snapping at a colleague over a minor issue. She immediately regrets it, realizing she’s been
+more irritable lately. “Maybe I’m just stressed,” she thinks. But how can she be sure? She wishes she had a way to keep
+track of her mood patterns over time.
 
 ## Lunch
 
-Over lunch, Zoe scrolls through social media and stumbles upon a post about mental health. The post mentions the benefits of journaling to track emotions and identify triggers. Zoe likes the idea but doesn’t feel like carrying around a notebook.
+Over lunch, Zoe scrolls through social media and stumbles upon a post about mental health. The post mentions the
+benefits of journaling to track emotions and identify triggers. Zoe likes the idea but doesn’t feel like carrying around
+a notebook.
 
 She jots down a quick note on her phone: “Track moods?” but quickly gets distracted by another notification.
 
 ## Afternoon
 
-While walking home, Zoe remembers a conversation with a friend who uses an app to log their daily emotions. “It helps me spot patterns,” the friend had said. Zoe thinks about downloading an app but feels overwhelmed by the endless options. “I just need something simple,” she mutters.
+While walking home, Zoe remembers a conversation with a friend who uses an app to log their daily emotions. “It helps me
+spot patterns,” the friend had said. Zoe thinks about downloading an app but feels overwhelmed by the endless options.
+“I just need something simple,” she mutters.
 
 ## Evening
 
-At home, Zoe reflects on her day. She wonders if her irritability earlier was due to poor sleep or if it’s been building up over time. She starts to write it down in her notebook but gets stuck on what to say. “How do I even begin tracking this?” she wonders.
+At home, Zoe reflects on her day. She wonders if her irritability earlier was due to poor sleep or if it’s been building
+up over time. She starts to write it down in her notebook but gets stuck on what to say. “How do I even begin tracking
+this?” she wonders.
 
 She closes the notebook, thinking, “There has to be a better way to track how I feel without overcomplicating things.”
